fastapi_missing/requirements.txt
# requirements.txt для missing_data_service
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fastapi==0.112.2
fonttools==4.54.1
h11==0.14.0
httpx==0.27.2
idna==3.8
jinja2==3.1.4
kiwisolver==1.4.7
markupsafe==2.1.5
matplotlib==3.9.2
missingno==0.5.2
numpy==2.1.3
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
pyyaml==6.0.2
redis==5.2.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
statsmodels==0.14.2
threadpoolctl==3.5.0
typing-extensions==4.12.2
tzdata==2024.2
uvicorn==0.30.0
watchfiles==0.24.0
websockets==13.0.1